requirements.txt
streamlit
requests
firebase_admin
python-dotenv
base64